automatic-portrait-tf-master/requirements.txt
Pillow
requests
tensorflow-gpu==1.6
numpy
scipy
